fix: move writing-style sample helper out of the server-actions file
Next.js treats every runtime export from a top-level "use server" file
as a callable Server Action; a synchronous helper export there fails
the build (caught by check-server-action-exports.js / CI). Moved the
helper into its own module alongside the AI call it feeds.
